refactor : 재정의 메소드 반영, 테스트 검증 로직 수정
